test(cursor): extract pagination and fetching tests to stay under file length limit
